requirements.txt
viam-sdk
transformers
pillow
torch
timm
einops